feat: add tests for llamastack lmeval provider
