+++ tests/model_explainability/constants.py
@@ -1,0 +1,1 @@
+MNT_MODELS: str = "/mnt/models"
