+++ tests/model_explainability/guardrails/test_llamastack_fms_provider.py
@@ -7,7 +7,7 @@ from tests.model_explainability.guardrails.constants import (
     BUILTIN_DETECTOR_CONFIG,
     PROMPT_WITH_PII,
 )
-from tests.model_explainability.guardrails.test_guardrails import MNT_MODELS
+from tests.model_explainability.constants import MNT_MODELS
 from utilities.constants import MinIo
 
 LOGGER = get_logger(name=__name__)
@@ -15,7 +15,8 @@ PII_REGEX_SHIELD_ID = "regex"
 
 
 @pytest.mark.parametrize(
-    "model_namespace, minio_pod, minio_data_connection, orchestrator_config, guardrails_orchestrator",
+    "model_namespace, minio_pod, minio_data_connection, "
+    "orchestrator_config, guardrails_orchestrator, llamastack_distribution",
     [
         pytest.param(
             {"name": "test-guardrails-lls"},
@@ -30,10 +31,13 @@ PII_REGEX_SHIELD_ID = "regex"
                 },
             },
             {"enable_built_in_detectors": True, "enable_guardrails_gateway": False},
+            {"guardrails_orchestrator_route_fixture": "guardrails_orchestrator_route"},
         )
     ],
     indirect=True,
 )
+@pytest.mark.rawdeployment
+@pytest.mark.usefixtures("orchestrator_config", "guardrails_orchestrator")
 class TestLlamaStackFMSGuardrailsProvider:
     """
     Adds basic tests for the LlamaStack FMS Guardrails provider.
@@ -47,11 +51,11 @@ class TestLlamaStackFMSGuardrailsProvider:
     4. Test a basic detection (PII) by using run_shields
     """
 
-    def test_fms_guardrails_register_model(self, qwen_isvc, llamastack_client_trustyai):
+    def test_fms_guardrails_register_model(self, qwen_isvc, llamastack_client):
         provider_id = "vllm-inference"
         model_type = "llm"
-        llamastack_client_trustyai.models.register(provider_id=provider_id, model_type=model_type, model_id=MNT_MODELS)
-        models = llamastack_client_trustyai.models.list()
+        llamastack_client.models.register(provider_id=provider_id, model_type=model_type, model_id=MNT_MODELS)
+        models = llamastack_client.models.list()
 
         # We only need to check the first model;
         # second is a granite embedding model present by default
@@ -60,8 +64,8 @@ class TestLlamaStackFMSGuardrailsProvider:
         assert models[0].provider_id == "vllm-inference"
         assert models[0].model_type == "llm"
 
-    def test_fms_guardrails_inference(self, minio_pod, qwen_isvc, llamastack_client_trustyai):
-        chat_completion_response = llamastack_client_trustyai.inference.chat_completion(
+    def test_fms_guardrails_inference(self, minio_pod, qwen_isvc, llamastack_client):
+        chat_completion_response = llamastack_client.inference.chat_completion(
             messages=[
                 {"role": "system", "content": "You are a friendly assistant."},
                 {"role": "user", "content": "Only respond with ack"},
@@ -72,7 +76,7 @@ class TestLlamaStackFMSGuardrailsProvider:
         assert chat_completion_response.completion_message.content != ""
 
     def test_fms_guardrails_register_shield(
-        self, current_client_token, qwen_isvc, patched_llamastack_deployment_tls_certs, llamastack_client_trustyai
+        self, current_client_token, qwen_isvc, patched_llamastack_deployment_tls_certs, llamastack_client
     ):
         trustyai_fms_provider_id = "trustyai_fms"
         shield_params = {
@@ -83,22 +87,22 @@ class TestLlamaStackFMSGuardrailsProvider:
             "verify_ssl": True,
             "ssl_cert_path": "/etc/llama/certs/orch-certificate.crt",
         }
-        llamastack_client_trustyai.shields.register(
+        llamastack_client.shields.register(
             shield_id=PII_REGEX_SHIELD_ID,
             provider_shield_id=PII_REGEX_SHIELD_ID,
             provider_id=trustyai_fms_provider_id,
             params=shield_params,
             timeout=120,
         )
-        shields = llamastack_client_trustyai.shields.list()
+        shields = llamastack_client.shields.list()
 
         assert len(shields) == 1
         assert shields[0].identifier == PII_REGEX_SHIELD_ID
         assert shields[0].provider_id == trustyai_fms_provider_id
         assert shields[0].params == shield_params
 
-    def test_fms_guardrails_run_shield(self, llamastack_client_trustyai):
-        run_shields_response = llamastack_client_trustyai.safety.run_shield(
+    def test_fms_guardrails_run_shield(self, llamastack_client):
+        run_shields_response = llamastack_client.safety.run_shield(
             shield_id=PII_REGEX_SHIELD_ID,
             messages=[
                 {
